fix: 게임 만들기 기능 버그 수정
