--- src/main/resources/templates/exception.hbs
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Klay's Chess Game</title>
-    <link rel="stylesheet" href="/style/style.css">
-</head>
-<body>
-    <div class="header center-object">
-        <a class="center-object" href ="/"><h1 class="main-message">초간단 체스게임♞(홈으로)</h1></a>
-    </div>
-    <h1 class="main-message center-object">
-        {{exception}}
-    </h1>
-    <div class="exception-buttons center-object">
-            <button type="button" class="button referrer-start-button">
-                새 게임 시작하기
-            </button>
-            <button type="button" class="button referrer-button">
-                돌아가기
-            </button>
-    </div>
-</body>
-<script src="/script/chess.js"></script>
-</html>
