+++ src/main/resources/templates/index.hbs
@@ -34,11 +34,11 @@
     </div>
     </br>
     <div class="center-object">
-        <form action="/board" method="post" >
+        <form action="/game" method="post" >
             <label for="name">새 게임 만들기: </label>
             <input type="text" name="name" id="name" value="이름" required>
             <input type="password" name="password" id="password" value="password" required>
-            <input type="submit">
+            <button id='create-game-button' type="submit">만들기!</button>
         </form>
     </div>
     <script src="/script/index.js"></script>
